tests/wpt/web-platform-tests/webmessaging/Channel_postMessage_DataCloneErr.htm

<!DOCTYPE html>
<html>
<head>
<title> postMessage() with a host object raises DataCloneError </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>

    var description = "Throw a DataCloneError when a host object (e.g. a DOM node) is used with postMessage.";

    test(function()
    {
        var channel = new MessageChannel();
        channel.port1.start();

        assert_throws_dom("DATA_CLONE_ERR", function()
        {
            channel.port1.postMessage(navigator);
        });
    }, description);
</script>
</body>
</html>
